+++ src/ImageProcessor/Samplers/Resamplers/MitchellNetravaliResampler.cs
@@ -12,13 +12,13 @@ namespace ImageProcessor.Samplers
     public class MitchellNetravaliResampler : IResampler
     {
         /// <inheritdoc/>
-        public double Radius => 2;
+        public float Radius => 2;
 
         /// <inheritdoc/>
-        public double GetValue(double x)
+        public float GetValue(float x)
         {
-            const double B = 1 / 3d;
-            const double C = 1 / 3d;
+            const float B = 1 / 3f;
+            const float C = 1 / 3f;
 
             return ImageMaths.GetBcValue(x, B, C);
         }
